Convert validate_unique_base_path from make to rake
The idea is to move everything to rake, to avoid the cost
of having to spin up ruby many times over (which is very slow)
