--- bin/ensure_example_base_paths_unique
@@ -1,37 +1,0 @@
-#!/usr/bin/env ruby
-
-$LOAD_PATH << File.dirname(__FILE__) + "/../lib"
-require 'json'
-
-def usage
-  $stderr.puts "USAGE: #{File.basename(__FILE__)} <path/to/examples/>"
-  $stderr.puts
-  $stderr.puts "Checks that all examples have a unique base path"
-end
-
-if ARGV.count == 0
-  usage
-  exit(1)
-end
-
-def base_path(example_file)
-  JSON.parse(File.read(example_file))['base_path']
-end
-
-grouped = ARGV.group_by {|file| base_path(file) }
-
-duplicates = grouped.select {|base_path, group| group.count > 1 }
-
-if duplicates.any?
-  puts "#{duplicates.count} duplicate(s) found:"
-
-  duplicates.each do |base_path, group|
-    group.each do |filename|
-      puts "  #{filename}"
-    end
-  end
-  exit(1)
-else
-  puts "OK: All example base_paths are unique"
-  exit(0)
-end
